additional status information output on serial
Outputting current millis as well as the lastStateChangeMillis
parameter. This way troubleshooting any problems with
shortCycleProtection() will be easier. The millis output is not the
exact count that was used by the function, but the difference should
not be significant in the vast majority of cases. Want all serial
output to be only in the main loop for now rather than scattered in the
individual functions.