power off on unrecognized mode
If mode is unrecognized, then power off. This prevents the system from
running forever if system was already running and then mode changed to
an unrecognized mode. The purpose is to add a mode 9 which is off;
however, it stood to reason that any number for which logic was not
written should be off.